requirements.txt
psutil
speedtest-cli
telebot
